requirements.txt

streamlit
fpdf
requests
beautifulsoup4
Pillow
